+++ Changelog.md
@@ -1,5 +1,8 @@
 Lilu Changelog
 ==============
+#### v1.4.1
+- Made applyLookupPatch support kernel patches by passsing null kext
+
 #### v1.4.0
 - Fixed mishandling user patches process list after processKernel API call
 - Fixed extra I/O in user patcher even when no patches were needed